backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
psycopg2-binary==2.9.9
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.13.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0
